+++ smartcontract/programs/doublezero-serviceability/src/processors/allowlist/device/add.rs
@@ -5,7 +5,8 @@ use crate::{
     globalstate::{globalstate_get, globalstate_write_with_realloc},
     pda::*,
 };
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 #[cfg(test)]
 use solana_program::msg;
 use solana_program::{
@@ -15,7 +16,7 @@ use solana_program::{
     pubkey::Pubkey,
 };
 
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct AddDeviceAllowlistArgs {
     pub pubkey: Pubkey,
 }