+++ smartcontract/programs/doublezero-serviceability/src/processors/device/interface/unlink.rs
@@ -1,7 +1,8 @@
 use core::fmt;
 
 use crate::{error::DoubleZeroError, globalstate::globalstate_get, helper::*, state::device::*};
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 use doublezero_program_common::types::NetworkV4;
 #[cfg(test)]
 use solana_program::msg;
@@ -12,7 +13,7 @@ use solana_program::{
     pubkey::Pubkey,
 };
 
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct DeviceInterfaceUnlinkArgs {
     pub name: String,
 }