+++ smartcontract/programs/doublezero-serviceability/src/processors/link/closeaccount.rs
@@ -4,7 +4,8 @@ use crate::{
     helper::*,
     state::{contributor::Contributor, device::*, link::*},
 };
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 use doublezero_program_common::types::NetworkV4;
 #[cfg(test)]
 use solana_program::msg;
@@ -16,7 +17,7 @@ use solana_program::{
 };
 use std::fmt;
 
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct LinkCloseAccountArgs {}
 
 impl fmt::Debug for LinkCloseAccountArgs {