+++ smartcontract/programs/doublezero-serviceability/src/processors/link/resume.rs
@@ -4,7 +4,8 @@ use crate::{
     helper::*,
     state::{contributor::Contributor, link::*},
 };
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 use core::fmt;
 #[cfg(test)]
 use solana_program::msg;
@@ -13,7 +14,7 @@ use solana_program::{
     entrypoint::ProgramResult,
     pubkey::Pubkey,
 };
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct LinkResumeArgs {}
 
 impl fmt::Debug for LinkResumeArgs {