+++ smartcontract/programs/doublezero-serviceability/src/processors/multicastgroup/create.rs
@@ -5,7 +5,8 @@ use crate::{
     pda::get_multicastgroup_pda,
     state::{accounttype::AccountType, multicastgroup::*},
 };
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 use doublezero_program_common::validate_account_code;
 use solana_program::{
     account_info::{next_account_info, AccountInfo},
@@ -18,7 +19,7 @@ use std::fmt;
 #[cfg(test)]
 use solana_program::msg;
 
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct MulticastGroupCreateArgs {
     pub index: u128,
     pub bump_seed: u8,