+++ smartcontract/programs/doublezero-serviceability/src/processors/multicastgroup/suspend.rs
@@ -3,7 +3,8 @@ use core::fmt;
 use crate::{
     error::DoubleZeroError, globalstate::globalstate_get, helper::*, state::multicastgroup::*,
 };
-use borsh::{BorshDeserialize, BorshSerialize};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
 #[cfg(test)]
 use solana_program::msg;
 use solana_program::{
@@ -12,7 +13,7 @@ use solana_program::{
     pubkey::Pubkey,
 };
 
-#[derive(BorshSerialize, BorshDeserialize, PartialEq, Clone, Default)]
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
 pub struct MulticastGroupSuspendArgs {}
 
 impl fmt::Debug for MulticastGroupSuspendArgs {